updated my code after I checked the solution. I did, forget that I changed the font style of "two" and so i removed that line as well. I appreciate the notes within the solution.css file. I didn't see the similarity between "three" and "four" for the font size. It is helpful to see some other ways to write the code.
Mine would've worked (other than "two" being in Time-New-Roman);
but the solution is definitely cleaner and more
stream-lined.

Oh, and i had originally edited the link on the html file; I
added a "./" before the style.css; but the solution doesn't have that, so
I figure they have a reason for it. and I will leave it alone in the
future.
I just remember, earlier within TOP, they said "./" at the start of href links
would be a safe bet.